+++ src/Switch.h
@@ -8,11 +8,11 @@ extern "C"
     extern void Switch_init();
     extern bool Switch_read();
     extern bool get_bmcu_selected();
-    extern void set_bmcu_selected(int selected);
-    extern int get_filament_map_to(int num);
-    extern bool check_bmcu_num(int bmcu);
-    extern int get_current_bmcu_num();
-    extern std::pair<int, int> get_bmcu_and_channel(int number);
+    extern void set_bmcu_selected(uint8_t selected);
+    extern uint8_t get_filament_map_to(uint8_t num);
+    extern bool check_bmcu_num(uint8_t bmcu);
+    extern uint8_t get_current_bmcu_num();
+    extern std::pair<uint8_t, uint8_t> get_bmcu_and_channel(uint8_t number);
     extern bool Switch_set_filament(unsigned char *buf, int length, uint8_t AMS_num, uint8_t read_num);
     extern void Switch_set_need_to_save();
     extern bool Switch_need_to_save();
